Crate google-gamesmanagement1_management

This documentation was generated from Games Management crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

achievements

reset, reset all, reset all for all players, reset for all players and reset multiple for all players

applications

list hidden

events

reset, reset all, reset all for all players, reset for all players and reset multiple for all players

players

hide and unhide

quests

reset, reset all, reset all for all players, reset for all players and reset multiple for all players

rooms

reset and reset for all players

scores

reset, reset all, reset all for all players, reset for all players and reset multiple for all players

turn based matches

reset and reset for all players

Everything else about the Games Management v1_management API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.players().hide(...).doit()
let r = hub.players().unhide(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-gamesmanagement1_management = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate serde;
extern crate "google-gamesmanagement1_management" as gamesmanagement1_management;
use gamesmanagement1_management::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use gamesmanagement1_management::GamesManagement;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = GamesManagement::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.players().hide("applicationId", "playerId")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::JsonDecodeError(err) => println!("Json failed to decode: {:?}", err),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Uploads and Downlods

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

AchievementMethodsBuilder
A builder providing access to all methods supported on achievement resources. It is not used directly, but through the GamesManagement hub.

AchievementResetAllCallBuilder
Resets all achievements for the currently authenticated player for your application. This method is only accessible to whitelisted tester accounts for your application.

AchievementResetAllForAllPlayerCallBuilder
Resets all draft achievements for all players. This method is only available to user accounts for your developer console.

AchievementResetAllResponse
This is a JSON template for achievement reset all response.

AchievementResetCallBuilder
Resets the achievement with the given ID for the currently authenticated player. This method is only accessible to whitelisted tester accounts for your application.

AchievementResetForAllPlayerCallBuilder
Resets the achievement with the given ID for all players. This method is only available to user accounts for your developer console. Only draft achievements can be reset.

AchievementResetMultipleForAllPlayerCallBuilder
Resets achievements with the given IDs for all players. This method is only available to user accounts for your developer console. Only draft achievements may be reset.

AchievementResetMultipleForAllRequest
This is a JSON template for multiple achievements reset all request.

AchievementResetResponse
This is a JSON template for an achievement reset response.

ApplicationListHiddenCallBuilder
Get the list of players hidden from the given application. This method is only available to user accounts for your developer console.

ApplicationMethodsBuilder
A builder providing access to all methods supported on application resources. It is not used directly, but through the GamesManagement hub.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

EventMethodsBuilder
A builder providing access to all methods supported on event resources. It is not used directly, but through the GamesManagement hub.

EventResetAllCallBuilder
Resets all player progress on all events for the currently authenticated player. This method is only accessible to whitelisted tester accounts for your application. All quests for this player will also be reset.

EventResetAllForAllPlayerCallBuilder
Resets all draft events for all players. This method is only available to user accounts for your developer console. All quests that use any of these events will also be reset.

EventResetCallBuilder
Resets all player progress on the event with the given ID for the currently authenticated player. This method is only accessible to whitelisted tester accounts for your application. All quests for this player that use the event will also be reset.

EventResetForAllPlayerCallBuilder
Resets the event with the given ID for all players. This method is only available to user accounts for your developer console. Only draft events can be reset. All quests that use the event will also be reset.

EventResetMultipleForAllPlayerCallBuilder
Resets events with the given IDs for all players. This method is only available to user accounts for your developer console. Only draft events may be reset. All quests that use any of the events will also be reset.

EventsResetMultipleForAllRequest
This is a JSON template for multiple events reset all request.

GamesManagement
Central instance to access all GamesManagement related resource activities

GamesPlayedResource
This is a JSON template for metadata about a player playing a game with the currently authenticated user.

GamesPlayerExperienceInfoResource
This is a JSON template for 1P/3P metadata about the player's experience.

GamesPlayerLevelResource
This is a JSON template for 1P/3P metadata about a user's level.

HiddenPlayer
This is a JSON template for the HiddenPlayer resource.

HiddenPlayerList
This is a JSON template for a list of hidden players.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Player
This is a JSON template for a Player resource.

PlayerHideCallBuilder
Hide the given player's leaderboard scores from the given application. This method is only available to user accounts for your developer console.

PlayerMethodsBuilder
A builder providing access to all methods supported on player resources. It is not used directly, but through the GamesManagement hub.

PlayerName
An object representation of the individual components of the player's name. For some players, these fields may not be present.

PlayerScoreResetAllResponse
This is a JSON template for a list of leaderboard reset resources.

PlayerScoreResetResponse
This is a JSON template for a list of reset leaderboard entry resources.

PlayerUnhideCallBuilder
Unhide the given player's leaderboard scores from the given application. This method is only available to user accounts for your developer console.

QuestMethodsBuilder
A builder providing access to all methods supported on quest resources. It is not used directly, but through the GamesManagement hub.

QuestResetAllCallBuilder
Resets all player progress on all quests for the currently authenticated player. This method is only accessible to whitelisted tester accounts for your application.

QuestResetAllForAllPlayerCallBuilder
Resets all draft quests for all players. This method is only available to user accounts for your developer console.

QuestResetCallBuilder
Resets all player progress on the quest with the given ID for the currently authenticated player. This method is only accessible to whitelisted tester accounts for your application.

QuestResetForAllPlayerCallBuilder
Resets all player progress on the quest with the given ID for all players. This method is only available to user accounts for your developer console. Only draft quests can be reset.

QuestResetMultipleForAllPlayerCallBuilder
Resets quests with the given IDs for all players. This method is only available to user accounts for your developer console. Only draft quests may be reset.

QuestsResetMultipleForAllRequest
This is a JSON template for multiple quests reset all request.

RoomMethodsBuilder
A builder providing access to all methods supported on room resources. It is not used directly, but through the GamesManagement hub.

RoomResetCallBuilder
Reset all rooms for the currently authenticated player for your application. This method is only accessible to whitelisted tester accounts for your application.

RoomResetForAllPlayerCallBuilder
Deletes rooms where the only room participants are from whitelisted tester accounts for your application. This method is only available to user accounts for your developer console.

ScoreMethodsBuilder
A builder providing access to all methods supported on score resources. It is not used directly, but through the GamesManagement hub.

ScoreResetAllCallBuilder
Resets all scores for all leaderboards for the currently authenticated players. This method is only accessible to whitelisted tester accounts for your application.

ScoreResetAllForAllPlayerCallBuilder
Resets scores for all draft leaderboards for all players. This method is only available to user accounts for your developer console.

ScoreResetCallBuilder
Resets scores for the leaderboard with the given ID for the currently authenticated player. This method is only accessible to whitelisted tester accounts for your application.

ScoreResetForAllPlayerCallBuilder
Resets scores for the leaderboard with the given ID for all players. This method is only available to user accounts for your developer console. Only draft leaderboards can be reset.

ScoreResetMultipleForAllPlayerCallBuilder
Resets scores for the leaderboards with the given IDs for all players. This method is only available to user accounts for your developer console. Only draft leaderboards may be reset.

ScoresResetMultipleForAllRequest
This is a JSON template for multiple scores reset all request.

TurnBasedMatcheMethodsBuilder
A builder providing access to all methods supported on turnBasedMatche resources. It is not used directly, but through the GamesManagement hub.

TurnBasedMatcheResetCallBuilder
Reset all turn-based match data for a user. This method is only accessible to whitelisted tester accounts for your application.

TurnBasedMatcheResetForAllPlayerCallBuilder
Deletes turn-based matches where the only match participants are from whitelisted tester accounts for your application. This method is only available to user accounts for your developer console.

Enums

Result
A universal result type used as return for all action method results.

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

ResponseResult
Identifies types which are used in API responses.

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.