Module google_gamesmanagement1_management::api

Structs

AchievementMethods
A builder providing access to all methods supported on achievement resources. It is not used directly, but through the GamesManagement hub.

AchievementResetAllCall
Resets all achievements for the currently authenticated player for your application. This method is only accessible to whitelisted tester accounts for your application.

AchievementResetAllForAllPlayerCall
Resets all draft achievements for all players. This method is only available to user accounts for your developer console.

AchievementResetAllResponse
Achievement reset all response.

AchievementResetCall
Resets the achievement with the given ID for the currently authenticated player. This method is only accessible to whitelisted tester accounts for your application.

AchievementResetForAllPlayerCall
Resets the achievement with the given ID for all players. This method is only available to user accounts for your developer console. Only draft achievements can be reset.

AchievementResetMultipleForAllPlayerCall
Resets achievements with the given IDs for all players. This method is only available to user accounts for your developer console. Only draft achievements may be reset.

AchievementResetMultipleForAllRequest
There is no detailed description.

AchievementResetResponse
An achievement reset response.

ApplicationListHiddenCall
Get the list of players hidden from the given application. This method is only available to user accounts for your developer console.

ApplicationMethods
A builder providing access to all methods supported on application resources. It is not used directly, but through the GamesManagement hub.

EventMethods
A builder providing access to all methods supported on event resources. It is not used directly, but through the GamesManagement hub.

EventResetAllCall
Resets all player progress on all events for the currently authenticated player. This method is only accessible to whitelisted tester accounts for your application.

EventResetAllForAllPlayerCall
Resets all draft events for all players. This method is only available to user accounts for your developer console.

EventResetCall
Resets all player progress on the event with the given ID for the currently authenticated player. This method is only accessible to whitelisted tester accounts for your application.

EventResetForAllPlayerCall
Resets the event with the given ID for all players. This method is only available to user accounts for your developer console. Only draft events can be reset.

EventResetMultipleForAllPlayerCall
Resets events with the given IDs for all players. This method is only available to user accounts for your developer console. Only draft events may be reset.

EventsResetMultipleForAllRequest
Multiple events reset all request.

GamesManagement
Central instance to access all GamesManagement related resource activities

GamesPlayerExperienceInfoResource
1P/3P metadata about the player's experience.

GamesPlayerLevelResource
1P/3P metadata about a user's level.

HiddenPlayer
The HiddenPlayer resource.

HiddenPlayerList
A list of hidden players.

Player
A Player resource.

PlayerHideCall
Hide the given player's leaderboard scores from the given application. This method is only available to user accounts for your developer console.

PlayerMethods
A builder providing access to all methods supported on player resources. It is not used directly, but through the GamesManagement hub.

PlayerName
An object representation of the individual components of the player's name. For some players, these fields may not be present.

PlayerScoreResetAllResponse
A list of leaderboard reset resources.

PlayerScoreResetResponse
A list of reset leaderboard entry resources.

PlayerUnhideCall
Unhide the given player's leaderboard scores from the given application. This method is only available to user accounts for your developer console.

ProfileSettings
Profile settings

ScoreMethods
A builder providing access to all methods supported on score resources. It is not used directly, but through the GamesManagement hub.

ScoreResetAllCall
Resets all scores for all leaderboards for the currently authenticated players. This method is only accessible to whitelisted tester accounts for your application.

ScoreResetAllForAllPlayerCall
Resets scores for all draft leaderboards for all players. This method is only available to user accounts for your developer console.

ScoreResetCall
Resets scores for the leaderboard with the given ID for the currently authenticated player. This method is only accessible to whitelisted tester accounts for your application.

ScoreResetForAllPlayerCall
Resets scores for the leaderboard with the given ID for all players. This method is only available to user accounts for your developer console. Only draft leaderboards can be reset.

ScoreResetMultipleForAllPlayerCall
Resets scores for the leaderboards with the given IDs for all players. This method is only available to user accounts for your developer console. Only draft leaderboards may be reset.

ScoresResetMultipleForAllRequest
There is no detailed description.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.