Struct google_gamesmanagement1_management::AchievementResetResponse

pub struct AchievementResetResponse {
    pub definition_id: Option<String>,
    pub kind: Option<String>,
    pub update_occurred: Option<bool>,
    pub current_state: Option<String>,
}

This is a JSON template for an achievement reset response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

reset achievements (response)

Fields

definition_id

The ID of an achievement for which player state has been updated.

kind

Uniquely identifies the type of this resource. Value is always the fixed string gamesManagement#achievementResetResponse.

update_occurred

Flag to indicate if the requested update actually occurred.

current_state

The current state of the achievement. This is the same as the initial state of the achievement. Possible values are:
- "HIDDEN"- Achievement is hidden. - "REVEALED" - Achievement is revealed. - "UNLOCKED" - Achievement is unlocked.