Struct google_gamesmanagement1_management::PlayerMethods

pub struct PlayerMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on player resources. It is not used directly, but through the GamesManagement hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_gamesmanagement1_management as gamesmanagement1_management;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use gamesmanagement1_management::GamesManagement;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = GamesManagement::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `hide(...)` and `unhide(...)`
// to build up your call.
let rb = hub.players();

Methods

impl<'a, C, A> PlayerMethods<'a, C, A>

fn hide(&self, application_id: &str, player_id: &str) -> PlayerHideCall<'a, C, A>

Create a builder to help you perform the following task:

Hide the given player's leaderboard scores from the given application. This method is only available to user accounts for your developer console.

Arguments

applicationId - The application ID from the Google Play developer console.

playerId - A player ID. A value of me may be used in place of the authenticated player's ID.

fn unhide(&self, application_id: &str, player_id: &str) -> PlayerUnhideCall<'a, C, A>

Create a builder to help you perform the following task:

Unhide the given player's leaderboard scores from the given application. This method is only available to user accounts for your developer console.

Arguments

applicationId - The application ID from the Google Play developer console.

playerId - A player ID. A value of me may be used in place of the authenticated player's ID.