Struct google_gamesmanagement1_management::QuestsResetMultipleForAllRequest

pub struct QuestsResetMultipleForAllRequest {
    pub quest_ids: Option<Vec<String>>,
    pub kind: Option<String>,
}

This is a JSON template for multiple quests reset all request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

reset multiple for all players quests (request)

Fields
quest_ids: Option<Vec<String>>

The IDs of quests to reset.

kind: Option<String>

Uniquely identifies the type of this resource. Value is always the fixed string gamesManagement#questsResetMultipleForAllRequest.