Struct google_gamesmanagement1_management::RoomMethods

pub struct RoomMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on room resources. It is not used directly, but through the GamesManagement hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_gamesmanagement1_management as gamesmanagement1_management;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use gamesmanagement1_management::GamesManagement;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = GamesManagement::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `reset(...)` and `reset_for_all_players(...)`
// to build up your call.
let rb = hub.rooms();

Methods

impl<'a, C, A> RoomMethods<'a, C, A>

fn reset_for_all_players(&self) -> RoomResetForAllPlayerCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes rooms where the only room participants are from whitelisted tester accounts for your application. This method is only available to user accounts for your developer console.

fn reset(&self) -> RoomResetCall<'a, C, A>

Create a builder to help you perform the following task:

Reset all rooms for the currently authenticated player for your application. This method is only accessible to whitelisted tester accounts for your application.