Struct google_gamesmanagement1_management::EventMethods

pub struct EventMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on event resources. It is not used directly, but through the GamesManagement hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_gamesmanagement1_management as gamesmanagement1_management;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use gamesmanagement1_management::GamesManagement;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = GamesManagement::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `reset(...)`, `reset_all(...)`, `reset_all_for_all_players(...)`, `reset_for_all_players(...)` and `reset_multiple_for_all_players(...)`
// to build up your call.
let rb = hub.events();

Methods

impl<'a, C, A> EventMethods<'a, C, A>

fn reset_all_for_all_players(&self) -> EventResetAllForAllPlayerCall<'a, C, A>

Create a builder to help you perform the following task:

Resets all draft events for all players. This method is only available to user accounts for your developer console. All quests that use any of these events will also be reset.

fn reset(&self, event_id: &str) -> EventResetCall<'a, C, A>

Create a builder to help you perform the following task:

Resets all player progress on the event with the given ID for the currently authenticated player. This method is only accessible to whitelisted tester accounts for your application. All quests for this player that use the event will also be reset.

Arguments

eventId - The ID of the event.

fn reset_all(&self) -> EventResetAllCall<'a, C, A>

Create a builder to help you perform the following task:

Resets all player progress on all events for the currently authenticated player. This method is only accessible to whitelisted tester accounts for your application. All quests for this player will also be reset.

fn reset_multiple_for_all_players(&self, request: EventsResetMultipleForAllRequest) -> EventResetMultipleForAllPlayerCall<'a, C, A>

Create a builder to help you perform the following task:

Resets events with the given IDs for all players. This method is only available to user accounts for your developer console. Only draft events may be reset. All quests that use any of the events will also be reset.

Arguments

request - No description provided.

fn reset_for_all_players(&self, event_id: &str) -> EventResetForAllPlayerCall<'a, C, A>

Create a builder to help you perform the following task:

Resets the event with the given ID for all players. This method is only available to user accounts for your developer console. Only draft events can be reset. All quests that use the event will also be reset.

Arguments

eventId - The ID of the event.